Initial commit
Much of killjoy's core logic is in place. It can read in a configuration
file, make use of non-regex monitoring rules to monitor units, and dump
events of interest to the console. However, it's far from complete.
Settings are not subject to semantic validation, regex monitoring rules
are unused, units are not contacted over the D-Bus, automated testing
needs a great deal of expansion, and so much more.