Update docstrings
Add missing docstrings to the public module killjoy::settings.

Change comments in killjoy::bus from docstring comments to regular
comments. The module is private, meaning that the comments won't be
compiled, and most markdown syntax (like links) will be ignored. Using
the `///` syntax is only useful if doctests are in use.