Type-cast fields on Killjoy settings objects
The `killjoy::settings::Settings` object is composed of types from the
standard library, such as strings and vectors. This makes it easy to
deserialize with serde, but the resulting object may contain values that
are syntactically correct but semantically incorrect. For example,
`settings.notifiers[i].bus_name` is a `String`, and it may be impossible
to cast this string to a `dbus::BusName` object.

This is especially problematic because killjoy may fail to make use of a
particular value until late in runtime. For example, killjoy might
notice an event of interest, decide to notify a notifier, and *then*
discover that a `bus_name` contains a malformed address.

Address this issue by making pairs of classes:

*   `SerdeNotification` and `Notification`
*   `SerdeRule` and `Rule`
*   `SerdeSettings` and `Settings`

Going forward, the former objects will continue to be defined with
simple types from the standard library, and will be easily deserialized.
Each of the former has a `to_*` method which creates an object of the
latter type, with properly cast types.

This should make it easy to discover syntactically valid but
semantically invalid values as soon as the JSON settings file is read
in, which is quite early in the start-up process. In addition, the
`killjoy settings validate` command can be updated to benefit from this
change.

As it stands, the conversion is not complete. Certain fields on the
`Settings` struct (and sub-structs) have not been converted to more
appropriate types. But this commit moves a significant amount in that
direction.