Make Expression/unit_name matching logic more OO
The `unit_name_matches_expression` function checks to see whether an
`Expression` object matches a `unit_name` string slice. This is a
violation of the OO design principle of bundling data and logic which
operates on that data.

Let the Expression/unit matching logic be implemented as methods on
`Expression`:

    // Given these objects…
    let unit_name = "foo.service";
    let expression = Expression::UnitName("foo.service".to_owned());

    // …here's now matching works without this commit…
    unit_name_matches_expression(&unit_name, &expression))

    // …and here's now matching works with this commit.
    expression.matches(&unit_name)