Refactor src/main.rs
This file is the entry point for the application, and it's mainly
concerned with figuring out which CLI arguments the user passed, and
handing off responsibility once that's been sorted out. Killjoy has a
simple CLI, but it's just complicated enough to make the CLI parsing
logic hard to understand.

Refactor this file by creating a collection of `handle_*_subcommand`
functions, where each handles the logic for just one subcommand. This
greatly reduces the amount of nested logic.