Use stricter type in killjoy::settings::Notifier
The `Notifier` struct has a `bus_name` attribute, which indicates which
D-Bus peer to contact. This attribute is a `String`, and using such a
permissive type is problematic for the reasons outlined in
b8a8ed808d6c58362f11abe498064e9007c30edb.

Make the `bus_name` field private, and guard access to it with `new()`
and `get_bus_name()`. Here's a brief look at the result:

    // OK.
    let valid_notifier = Notifier::new(
        "org.freedesktop.Notifications",
        BusType::Session,
    ).unwrap();

    // Panic!
    let invalid_notifier = Notifier::new(
        "org/freedesktop/Notifications",
        BusType::Session,
    ).unwrap();