Improve validation of the Settings struct
The `Settings` struct has the following declaration:

    pub struct Settings {
        pub notifiers: HashMap<String, Notifier>,
        pub rules: Vec<Rule>,
    }

Each of the `rules` may reference zero or more notifiers to contact when
the rule is triggered. An ideal way to do this would be with references
to the notifiers, like so:

    pub struct Rule {
        pub active_states: HashSet<ActiveState>,
        pub bus_type: BusType,
        pub expression: Expression,
        pub notifiers: Vec<&Notifier>, // :-)
    }

Unfortunately, rules currently reference the `notifiers` hashmap by
string key:

    pub struct Rule {
        pub active_states: HashSet<ActiveState>,
        pub bus_type: BusType,
        pub expression: Expression,
        pub notifiers: Vec<String>, // :-(
    }

It would be ideal to use references, but that forces an enormous number
of lifetime parameters to be annotated. As a hacky solution for now,
make the method responsible for creating new `Settings` objects verify
the validity of these keys.