Make killjoy contact notifiers
When killjoy notices an event of interest (e.g. foo.service changing to
the failed state), it should contact the notifiers listed in the
configuration file. Make killjoy do so, by fleshing out the
`BusWatcher.gen_on_change` method.

There are three aspects of the implementation worth noting.

First, the notifier's bus name is used to generate the notifier's path.
For example, if killjoy is told to contact com.example.MyNotifier1, it
will do so with the path /com/example/MyNotifier1. Abiding by this
pattern is a best practice, and killjoy is defining a new API, so it
gets to set the rules. But there's no escape hatch right now. There's no
way for a user to say "contact com.example.MyNotifier1 at path
/com/example/Foo."

Second, and more seriously, killjoy does not care about receiving
responses to the messages that it sends to notifiers. Although killjoy
doesn't really need to receive responses, expecting them could improve
usability. For example, imagine that a user configured killjoy to
contact com.example.MyTypo1. Because killjoy does not check for
responses after sending a message, it will not learn that this message
bounced, and it can't inform the user of their error (e.g. by printing a
message to stderr).

Third, the API is not documented anywhere. The implementation is the
specification. This is bad!