Send old unit state to notifier
Killjoy sends four fields to notifiers whenever an event of interest
occurs:

1.  unit name
2.  old state
3.  new state
4.  timestamp

"old state" is currently a placeholder string, "FIXME." This commit
makes killjoy send a more meaningful string, like "active."

In the case where killjoy does not know the old state of the unit at
hand, it will send the string "(unknown)." This is non-ideal. Notifiers
that wish to act on the old state, e.g. by changing how output looks
based on whether the old state is known, must check for this hard-coded
string.

It would be better to encode this information into the type system. Two
options spring to mind:

*   If D-Bus offers an "option" type, the old state field could be
    changed to this type.
*   The "old state" and "new state" fields could be merged, into a
    single "states" field. This field would be an array of states in
    some order (oldest to newest or vice versa).