Drop public API
Migrate the contents of `src/lib.rs` to other files, and then drop
`src/lib.rs`. Dropping that file causes doctests to stop running, so
migrate doctests to unit tests.

Make the documentation in `src/main.rs` be in sync with the contents of
the readme.

For reasoning behind this change, see #11.

Fix: #11