Cast argument to type earlier in runtime
When a user executes `killjoy settings validate path/to/settings.json`,
the last argument is currently represented as a string. This works, but
it's overly lax. Change the code base, so that the path is wrapped in a
`std::path::Path` object early at runtime. This stricter typing makes
the code more semantically rich and helps to catch certain types of
errors earlier during runtime.