Refactor integration tests
Make greater use of method chaining, and explicitly check exit codes
with `.code(code)` instead of calling `.success()` or `.failure()`. The
latter change improves test coverage. The `.failure()` method will
succeed regardless of whether an executable exits in a controlled way or
due to a panic, and asserting on exit codes helps to distinguish between
the two cases.