Convert unwrap() to expect()
Many calls to unwrap() or expect() can be eliminated by improving the
design of this application. But doing that takes a significant amount of
time and effort, and some calls will likely never be eliminated. (For
example, there's little point to eliminating such calls in unit tests.)
A very easy way to make the code more friendly is to ensure that panics,
when they occur, include a prose message explaining what went wrong.