Expand documentation
A sample settings file is now included in the documentation, so it's
been removed from `scripts/`.

Once killjoy receives a proper first release, `README.md` should be
drastically cut down, and `src/main.rs` should be the One True Source;
or vice versa.

It would be nice if the settings file in the documentation could be
automatically validated in some manner. Perhaps the `examples/`
directory could be useful?