Move documentation from source to readme
Documentation on docs.rs is showing the contents of the readme, not the
documentation in `src/main.rs`. This is surprising, as for the crates
I've looked at, documentation comes from `src/main.rs`.

I strongly suspect that this is so because killjoy is an application,
not a library. I suspect that applications' documentation comes from
their readme, whereas libraries' documentation comes from `src/main.rs`.

Move documentation from the source code to the readme.