Drop cargo cache in Travis CI builds
Cache handling consumes a significant amount of CI build time. For
example, on a recent run:

*   "adding /home/travis/.cargo to cache" consumed ~120s.
*   "store build cache" consumed ~140s.

Dropping the cache step drastically increases the time spent downloading
and compiling crates. The time spent on `scripts/travis-checks.sh` jumps
from ~20s to ~170s. This is still a net improvement of over 100s.