Better isolate killjoy from test host
Prepend `dbus-run-session` to every call to killjoy. This ensures that
killjoy doesn't accidentally connect to the test host's session D-Bus
instance, or communicate with any of its peers. The point of this change
is twofold:

*   Tests are less dependent upon the specific configuration of the host
    running tests.
*   Malfunctioning tests don't accidentally mess up the test host's
    environment.

Virtually all existing tests don't actually communicate with the session
bus, so this shouldn't have a practical impact. This change is a
precaution, not a solution to a current problem.

The only test which is excepted from this change is
`test_run_settings_success()`, which contacts systemd over the system
bus. It would be better if the test could operate like so:

1.  Create a temporary bus with dbus-run-session.
2.  Create a temporary systemd instance, and make it available on the
    temporary bus.
3.  Let killjoy connect to the temporary bus and communicate with the
    temporary systemd instance.

…but it's unclear to me how to do this. Work on this will have to wait
until a future time.

Also, change the unit names in the sample unit configurations to random
uuid4 values. This is another precaution against contacting real
services on the test host.